backend/requirements.txt
amqp==2.2.2
attrs==17.4.0
awsebcli==3.12.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
blessed==1.14.2
botocore==1.9.6
bs4==0.0.1
celery==4.1.0
cement==2.8.2
cffi==1.11.5
colorama==0.3.7
Django==2.0.2
django-celery-results==1.0.1
django-filter==1.1.0
djangorestframework==3.7.7
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
httpie==0.9.9
jmespath==0.9.3
kombu==4.1.0
lxml==4.2.0
mysqlclient==1.3.12
numpy==1.14.1
pathspec==0.5.5
pep8==1.7.1
pika==0.11.2
pluggy==0.6.0
py==1.5.2
pycparser==2.18
Pygments==2.2.0
pytest==3.4.1
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.9.1
scikit-learn==0.19.1
scipy==1.0.0
semantic-version==2.5.0
six==1.11.0
sklearn==0.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.47.0
